Decomposed Dashboard Greeting into separate file
