+++ ENGAGEHF/Dashboard/DashboardView.swift
@@ -11,39 +11,7 @@
 //  Created by Nick Riedman on 4/3/24.
 //
 
-import SpeziAccount
-import SpeziMockWebService
 import SwiftUI
-
-struct Greeting: View {
-    @State private var dateString: String?
-    
-    var body: some View {
-        HStack {
-            // Todo: replace world! with first name from account info
-            Text("Hello, world!")
-                .font(.title2)
-                .accessibilityLabel(Text("DASHBOARD_GREETING"))
-            Spacer()
-            Text(dateString ?? "No date")
-                .font(.title2)
-                .accessibilityLabel(Text("DASHBOARD_DATE"))
-        }
-        .padding()
-        .task {
-            getDateString()
-        }
-    }
-    
-    private func getDateString() {
-        let currentDate = Date()
-        let dateFormatter = DateFormatter()
-        
-        dateFormatter.dateStyle = .short
-        
-        dateString = dateFormatter.string(from: currentDate)
-    }
-}
 
 struct Dashboard: View {
     @Binding var presentingAccount: Bool
